Move error categories to source file
