+++ hpx/error_code.hpp
@@ -4,7 +4,7 @@
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-/// \file exception.hpp
+/// \file error_code.hpp
 
 #if !defined(HPX_ERROR_CODE_MAR_24_2008_0929AM)
 #define HPX_ERROR_CODE_MAR_24_2008_0929AM
@@ -19,56 +19,12 @@
 #include <stdexcept>
 #include <string>
 
-#include <hpx/config/warnings_prefix.hpp>
-
-#if !defined(BOOST_SYSTEM_NOEXCEPT)
-#define BOOST_SYSTEM_NOEXCEPT HPX_NOEXCEPT
-#endif
-
 ///////////////////////////////////////////////////////////////////////////////
 namespace hpx
 {
     /// \cond NODETAIL
     namespace detail
     {
-        class hpx_category : public boost::system::error_category
-        {
-        public:
-            const char* name() const BOOST_SYSTEM_NOEXCEPT
-            {
-                return "HPX";
-            }
-
-            std::string message(int value) const
-            {
-                if (value >= success && value < last_error)
-                    return std::string("HPX(") + error_names[value] + ")"; //-V108
-                if (value & system_error_flag)
-                    return std::string("HPX(system_error)");
-                return "HPX(unknown_error)";
-            }
-        };
-
-        struct lightweight_hpx_category : hpx_category {};
-
-        // this doesn't add any text to the exception what() message
-        class hpx_category_rethrow : public boost::system::error_category
-        {
-        public:
-            const char* name() const BOOST_SYSTEM_NOEXCEPT
-            {
-                return "";
-            }
-
-            std::string message(int) const HPX_NOEXCEPT
-            {
-                return "";
-            }
-        };
-
-        struct lightweight_hpx_category_rethrow : hpx_category_rethrow {};
-
-        ///////////////////////////////////////////////////////////////////////
         HPX_EXPORT boost::exception_ptr access_exception(error_code const&);
 
         ///////////////////////////////////////////////////////////////////////
@@ -87,44 +43,16 @@ namespace hpx
 
     ///////////////////////////////////////////////////////////////////////////
     /// \brief Returns generic HPX error category used for new errors.
-    inline boost::system::error_category const& get_hpx_category()
-    {
-        static detail::hpx_category instance;
-        return instance;
-    }
+    HPX_EXPORT boost::system::error_category const& get_hpx_category();
 
     /// \brief Returns generic HPX error category used for errors re-thrown
     ///        after the exception has been de-serialized.
-    inline boost::system::error_category const& get_hpx_rethrow_category()
-    {
-        static detail::hpx_category_rethrow instance;
-        return instance;
-    }
+    HPX_EXPORT boost::system::error_category const& get_hpx_rethrow_category();
 
     /// \cond NOINTERNAL
-    inline boost::system::error_category const&
-    get_lightweight_hpx_category()
-    {
-        static detail::lightweight_hpx_category instance;
-        return instance;
-    }
+    HPX_EXPORT boost::system::error_category const& get_lightweight_hpx_category();
 
-    inline boost::system::error_category const& get_hpx_category(throwmode mode)
-    {
-        switch(mode) {
-        case rethrow:
-            return get_hpx_rethrow_category();
-
-        case lightweight:
-        case lightweight_rethrow:
-            return get_lightweight_hpx_category();
-
-        case plain:
-        default:
-            break;
-        }
-        return get_hpx_category();
-    }
+    HPX_EXPORT boost::system::error_category const& get_hpx_category(throwmode mode);
 
     inline boost::system::error_code
     make_system_error_code(error e, throwmode mode = plain)
@@ -455,6 +383,5 @@ namespace hpx
 }
 
 #include <hpx/throw_exception.hpp>
-#include <hpx/config/warnings_suffix.hpp>
 
 #endif
